Feature/faisalramd/upload variant data

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/data/model/params/add/Product.kt
@@ -23,7 +23,7 @@ data class Product (
         var status: String = "",
         @SerializedName("stock")
         @Expose
-        var stock: Long = 0,
+        var stock: Int = 0,
         @SerializedName("pictures")
         @Expose
         var pictures: List<Picture> = emptyList()
